pyproject.toml
[project]
name = "timetable-remake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "alembic>=1.18.4",
    "celery>=5.6.2",
    "fastapi[standard]>=0.129.0",
    "ortools>=9.15.6755",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
]
